+++ contracts/mocks/MockFreeCollateral.sol
@@ -61,7 +61,7 @@ contract MockFreeCollateral is StorageLayoutV1 {
 
     function enableBitmapForAccount(
         address account,
-        uint256 currencyId,
+        uint16 currencyId,
         uint256 blockTime
     ) external {
         AccountContext memory accountContext = AccountContextHandler.getAccountContext(account);
